BIG refactoring
Basically, OOPified the program and made it actually work (BIG thanks to suitsense in https://stackoverflow.com/questions/57092430/how-to-call-script-with-python-in-the-same-directory-with-subprocesses ).

Also added some test for TDD things. It's not quite what you would see everywhere when doing TDD, as it's more of a tester of the whole thing rather than just a few methods of a class.

As a part of the refactoring, the code has been refactored according to PEP 8 guidelines. Well, mostly.

Now I guess it's time to write some documentation.